Channel and CRC32 Checksum
Channel is used to simulate transmission errors in the Physical Layer.
CRC32 checksum is added to the Link Layer header to detect corrupted frames.
if a frame is found to be corrupted upon arrival at the Link Layer, it is discarded silently
timeouts handle retransmissions of lost or corrupted frames.